requirements.txt
selenium>=4.0.0
pytest>=7.0.0
pytest-html>=3.1.1
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.23.0
pyautogui>=0.9.53